requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
idna==2.10
importlib-metadata==1.7.0
isort==5.3.2
Markdown==3.2.2
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
